requirements.txt
# FastAPI核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# 数据库相关
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL驱动

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 异步任务处理
celery==5.3.4
redis==5.0.1

# AI和机器学习
openai==1.3.0
numpy==2.2.6
pydub==0.25.1
torch==2.7.1
torchaudio==2.7.1

# HTTP客户端
httpx==0.25.2
requests==2.31.0

# 其他工具
python-dotenv==1.0.0
email-validator==2.1.0
aiofiles==24.1.0
tiktoken==0.6.0
